refactor useLoadEngineWithScene into systems
