+++ packages/client-core/src/world/Location.tsx
@@ -41,7 +41,7 @@ import LoadingView from '@ir-engine/ui/src/primitives/tailwind/LoadingView'
 import { StyledEngineProvider } from '@mui/material/styles'
 import { useTranslation } from 'react-i18next'
 import { NotificationService } from '../common/services/NotificationService'
-import { useLoadEngineWithScene, useNetwork } from '../components/World/EngineHooks'
+import { useNetwork } from '../components/World/EngineHooks'
 import { LocationService } from '../social/services/LocationService'
 import { LoadingUISystemState } from '../systems/LoadingUISystem'
 import { clientContextParams } from '../util/ClientContextState'
@@ -67,8 +67,6 @@ const LocationPage = ({ online }: Props) => {
 
   AuthService.useAPIListeners()
   LocationService.useLocationBanListeners()
-
-  useLoadEngineWithScene()
 
   useEffect(() => {
     if (!ready.value) return
